package change to dataOld in eclipse doorgevoerd
